Modified auth.service signin. New login endpoint authentiction
The auth.service signIn method was split into `signInUser` and
`genToken`. The signInUser asynchronously returns the user found
by the credentials, and genToken creates a token based off of the
given user. This was changed to support the http authentication
strategy, which required getting just the user, not the token.

Added the @AuthUser decorator. This takes the user from the request
object (`req.user`) and creates a user entity out of it. Specific
properties can be access from the decorator as well.

The endpoint `/auth/login` now uses HTTP authentication instead of
using just a post body. The post body was just for testing purposes.