frontend: adding overview components for each available model
Basically, there are four pages that will be used to view each model a
user can have. For example, there is a component called recipes-main
which provides an overview of all of the user's recipes. Each main
component will also have a component for viewing and modifying an
individual item. Each high-level component is routed.

Todo:
- Change toolbar color to primary
- Use title to redirect to dashboard