frontend: scaffolded profile view component
The profile view component shows user information. As of now, it
has a template user object that just shows basic user info in a
box to the left. When resource components are added, they will go
in the tab view on the right. Backgrounds are just to show the boxes
I am working with. They will be changed later, and will be on theme.