frontend: made recipe card component and modified recipes view
Made a new component to show an overview of a given recipe. It
shows an image of the recipe, its name, and the author's name.
I also added a grid-list layout to the recipes layout. However,
I am having issues getting uniform sizing for the cards. It may
be an issue with the size of the images, but still, I need to
look into it. I am considering using `@angular/flex-layout`
instead of `mat-grid-list`.

TODO:
- Add image column to the recipe model entity
- Remove mat-grid-list and use flex layout instead