frontend: starting a new frontend. Using angular material schematics
In this branched frontend, I am using more prebuilt schematics to make
prototyping the frontend easier. I am using the angular material
dashboard and navigation schematics. So far, the main app navigation
is set up to be resonsive. A baseline recipe dashboard is setup with a
few mock recipes. Their only issue is that the images in the cards are
not sized properly. I might just have to manually resize each photo
that comes in. Oh well.