frontend: mocked up pages for viewing recipes and user info
So far, have a dashboard for viewing recipes with recipe cards
that show just the basics and a page used for showing the indepth
info about a recipe. There is a specfic component for viewing an
ingredient list.

Have basic page for viewing user info. As of now, it just has info.
Nothing more.