frontend: made component for editing ingredients
I worked for a while to make a component that could be used as a custom
for control for editing ingredients. However, I had a bunch of issues
with the custom form control stuff. So, I am going to directly integrate
the component into the recipe edit page. This way, I can avoid the whole
issue of managing a custom form control.