frontend: recipe edit can now edit ingredients
Added a table in the `recipe-edit` component that the user can edit
ingredient lists with. A user can add new ingredients and modify
existing ones.

Going forward, it could be placed in a separate component, but that is a
problem for later. Plus, I believe it will only be used here.